+++ capsule/src/main/java/ca/allanwang/capsule/library/activities/ViewActivity.java
@@ -18,8 +18,11 @@ import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.view.ViewGroup;
 
+import org.greenrobot.eventbus.EventBus;
+
 import ca.allanwang.capsule.library.R;
 import ca.allanwang.capsule.library.custom.CapsuleCoordinatorLayout;
+import ca.allanwang.capsule.library.event.CClickEvent;
 import ca.allanwang.capsule.library.event.SnackbarEvent;
 import ca.allanwang.capsule.library.interfaces.CCollapseListener;
 import ca.allanwang.capsule.library.utils.AnimUtils;
@@ -36,6 +39,7 @@ abstract class ViewActivity extends PermissionActivity {
     protected CapsuleCoordinatorLayout cCoordinatorLayout;
     protected TabLayout cTabs;
     protected CollapsingToolbarLayout cCollapsingToolbarLayout;
+    protected boolean cToolbarClick = false;
 
     @Override
     public FloatingActionButton getFab() {
@@ -91,7 +95,7 @@ abstract class ViewActivity extends PermissionActivity {
      *
      * @return current fragment in view
      */
-    protected Fragment getCurrentFragment() {
+    public Fragment getCurrentFragment() {
         return getSupportFragmentManager().findFragmentById(getFragmentId());
     }
 
@@ -100,7 +104,7 @@ abstract class ViewActivity extends PermissionActivity {
      *
      * @return inner class holding capsulate methods
      */
-    protected Capsulate capsulate() {
+    public Capsulate capsulate() {
         return new Capsulate();
     }
 
@@ -142,7 +146,7 @@ abstract class ViewActivity extends PermissionActivity {
      * @param view view to add
      * @return view added
      */
-    protected <T extends View> T addCollapsingToolbarView(T view) {
+    public <T extends View> T addCollapsingToolbarView(T view) {
         if (cCollapsingToolbarLayout == null)
             throw new NullPointerException(sf(R.string.capsule_generic_not_set,
                     "CollapsingToolbarLayout"));
@@ -157,7 +161,7 @@ abstract class ViewActivity extends PermissionActivity {
      * @param layoutId id of layout
      * @return view created
      */
-    protected View addCollapsingToolbarView(@LayoutRes int layoutId) {
+    public View addCollapsingToolbarView(@LayoutRes int layoutId) {
         View view = getLayoutInflater().inflate(layoutId, null);
         return addCollapsingToolbarView(view);
     }
@@ -183,24 +187,27 @@ abstract class ViewActivity extends PermissionActivity {
      *
      * @param title String
      */
-    protected void setTitle(@Nullable String title) {
+    public void setTitle(@Nullable String title) {
         if (title == null) return;
         cToolbar.setTitle(title);
         if (cCollapsingToolbarLayout != null) cCollapsingToolbarLayout.setTitle(title);
     }
 
-    protected void snackbar(SnackbarEvent event) {
+    public void snackbar(SnackbarEvent event) {
         postEvent(event);
     }
 
     /**
      * Capsulate <p> Helps with initializing and managing other types of views
      */
-    protected class Capsulate {
+    public class Capsulate {
 
         public Capsulate toolbar(@IdRes int id) {
             cToolbar = (Toolbar) findViewById(id);
             setSupportActionBar(cToolbar);
+            cToolbar.setOnClickListener(v -> {
+                if (cToolbarClick) EventBus.getDefault().post(new CClickEvent(v));
+            });
             return this;
         }
 
@@ -264,6 +271,11 @@ abstract class ViewActivity extends PermissionActivity {
 
         public CustomizeToolbar setCollapseListener(@NonNull CCollapseListener listener) {
             cAppBarLayout.addOnOffsetChangedListener(listener);
+            return this;
+        }
+
+        public CustomizeToolbar withClickEvents(boolean enable) {
+            cToolbarClick = enable;
             return this;
         }