+++ capsule/src/main/java/ca/allanwang/capsule/library/event/CClickEvent.java
@@ -1,0 +1,17 @@
+package ca.allanwang.capsule.library.event;
+
+import android.view.View;
+
+/**
+ * Created by Allan Wang on 29/04/2017.
+ */
+
+public class CClickEvent {
+
+    public final View view;
+
+    public CClickEvent(View view) {
+        this.view = view;
+    }
+
+}