+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/SwipeRecyclerView.java
@@ -13,13 +13,14 @@ import android.util.AttributeSet;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
 import android.view.View;
+import android.view.animation.Interpolator;
 import android.widget.FrameLayout;
 
 import ca.allanwang.capsule.library.R;
 import ca.allanwang.capsule.library.logging.CLog;
 import ca.allanwang.capsule.library.swiperecyclerview.adapters.AnimationAdapter;
 import ca.allanwang.capsule.library.swiperecyclerview.animators.SlidingAnimator;
-import ca.allanwang.capsule.library.swiperecyclerview.interfaces.IScrolling;
+import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ILayoutManager;
 import ca.allanwang.capsule.library.swiperecyclerview.interfaces.ISwipeRecycler;
 import ca.allanwang.capsule.library.swiperecyclerview.managers.SGridLayoutManager;
 import ca.allanwang.capsule.library.swiperecyclerview.managers.SLinearLayoutManager;
@@ -174,10 +175,49 @@ public class SwipeRecyclerView extends FrameLayout implements SwipeRefreshBase.I
     }
 
     public SwipeRecyclerView setScrollingEnabled(boolean flag) {
-        if (!(mRecycler.getLayoutManager() instanceof IScrolling))
-            CLog.e("Recycler Layout Manager does not have custom IScrolling toggle");
+        if (!(mRecycler.getLayoutManager() instanceof ILayoutManager))
+            CLog.e("Recycler Layout Manager does not have custom ILayoutManager toggle");
         else
-            ((IScrolling) mRecycler.getLayoutManager()).setScrollEnabled(flag);
+            ((ILayoutManager) mRecycler.getLayoutManager()).setScrollEnabled(flag);
+        return this;
+    }
+
+    public SwipeRecyclerView scrollToPosition(int x, int y) {
+        mRecycler.scrollTo(x, y);
+        return this;
+    }
+
+    public SwipeRecyclerView smoothScrollToPosition(int position) {
+        mRecycler.smoothScrollToPosition(position);
+        return this;
+    }
+
+    /**
+     * If ILayoutManager is used, tries to set the duration in the Layout Manager for the next scroll
+     *
+     * @param position   to scroll to
+     * @param scrollSpeed callback to get scroll time
+     * @return this
+     */
+    public SwipeRecyclerView smoothScrollToPosition(int position, ILayoutManager.ScrollSpeed scrollSpeed) {
+        if (getLayoutManager() instanceof ILayoutManager)
+            ((ILayoutManager) getLayoutManager()).setSmoothScrollDuration(scrollSpeed);
+        else
+            CLog.d("LayoutManager is not an instance of ILayoutManager; cannot set smooth scroll duration");
+        return smoothScrollToPosition(position);
+    }
+
+    public SwipeRecyclerView smoothScrollBy(int dx, int dy) {
+        return smoothScrollBy(dx, dy, null);
+    }
+
+    public SwipeRecyclerView smoothScrollBy(int dx, int dy, Interpolator interpolator) {
+        mRecycler.smoothScrollBy(dx, dy, interpolator);
+        return this;
+    }
+
+    public SwipeRecyclerView stopScroll() {
+        mRecycler.stopScroll();
         return this;
     }
 
@@ -275,5 +315,13 @@ public class SwipeRecyclerView extends FrameLayout implements SwipeRefreshBase.I
      */
     public SwipeRefreshLayout getSwipeRefreshLayout() {
         return mSwipe;
+    }
+
+    public LinearLayoutManager getLayoutManager() {
+        return (LinearLayoutManager) mRecycler.getLayoutManager();
+    }
+
+    public boolean isFirstItemCompletelyVisible() {
+        return getLayoutManager().findFirstCompletelyVisibleItemPosition() == 0;
     }
 }