+++ capsule/src/main/java/ca/allanwang/capsule/library/swiperecyclerview/interfaces/IScrolling.java
@@ -1,9 +1,0 @@
-package ca.allanwang.capsule.library.swiperecyclerview.interfaces;
-
-/**
- * Created by Allan Wang on 2017-03-09.
- */
-
-public interface IScrolling {
-    void setScrollEnabled(boolean flag);
-}